requirements.txt
boto3==1.28.40
Pillow==10.0.0